backend/requirements.txt
# Web framework
fastapi>=0.115.0
uvicorn[standard]>=0.34.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.15.0

# Auth
python-jose[cryptography]>=3.4.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.20

# AI / RAG
langchain>=0.3.0
langchain-google-genai>=2.1.0
langchain-community>=0.3.0
langchain-chroma>=0.2.0
chromadb>=1.0.0
pypdf>=5.0.0

# Utilities
python-dotenv>=1.0.0
pydantic[email]>=2.10.0
